rename main package and move HALConfig to app
