+++ app/src/androidTest/java/br/com/stone/posandroid/hal/demo/bc/base/AutoLoadTableTest.kt
@@ -1,7 +1,6 @@
 package br.com.stone.posandroid.hal.demo.bc.base
 
 import androidx.test.platform.app.InstrumentationRegistry
-import br.com.stone.hal_demo.R
 import br.com.stone.posandroid.hal.api.Properties
 import br.com.stone.posandroid.hal.api.bc.Pinpad
 import br.com.stone.posandroid.hal.api.bc.PinpadCallbacks
@@ -10,6 +9,7 @@ import br.com.stone.posandroid.hal.api.bc.PinpadResult.Companion.TLI
 import br.com.stone.posandroid.hal.api.bc.constants.ResultCode.Companion.PP_OK
 import br.com.stone.posandroid.hal.api.bc.constants.RuntimeProperties
 import br.com.stone.posandroid.hal.demo.HALConfig
+import br.com.stone.posandroid.hal.demo.R
 import br.com.stone.posandroid.hal.mock.bc.PinpadStub.Companion.CombinedResult
 import io.mockk.mockk
 import org.junit.After
