+++ app/src/androidTest/java/br/com/stone/posandroid/hal/demo/bc/base/AutoOpenCloseTest.kt
@@ -2,7 +2,6 @@ package br.com.stone.posandroid.hal.demo.bc.base
 
 import androidx.test.internal.runner.junit4.AndroidJUnit4ClassRunner
 import androidx.test.platform.app.InstrumentationRegistry
-import br.com.stone.hal_demo.R
 import br.com.stone.posandroid.hal.api.Properties.KEY_CONTEXT
 import br.com.stone.posandroid.hal.api.Properties.RESULTS_FILE_KEY
 import br.com.stone.posandroid.hal.api.Properties.RESULTS_KEY
@@ -15,6 +14,7 @@ import br.com.stone.posandroid.hal.api.bc.PinpadResult.Companion.OPN
 import br.com.stone.posandroid.hal.api.bc.constants.ResultCode.Companion.PP_OK
 import br.com.stone.posandroid.hal.api.bc.constants.RuntimeProperties
 import br.com.stone.posandroid.hal.demo.HALConfig.deviceProvider
+import br.com.stone.posandroid.hal.demo.R
 import br.com.stone.posandroid.hal.mock.bc.PinpadStub.Companion.CombinedResult
 import io.mockk.mockk
 import org.junit.After
