+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="br.com.stone.hal_demo">
+    package="br.com.stone.posandroid.hal.demo">
 
     <uses-permission android:name="android.permission.MODIFY_PHONE_STATE" />
