+++ app/src/main/java/br/com/stone/posandroid/hal/demo/HALConfig.kt
@@ -1,0 +1,10 @@
+package br.com.stone.posandroid.hal.demo
+
+import br.com.stone.posandroid.hal.api.provider.AutoProvider
+import br.com.stone.posandroid.hal.api.provider.DeviceProvider
+import br.com.stone.posandroid.hal.mock.provider.MockProvider
+
+object HALConfig {
+    val deviceProvider: DeviceProvider = AutoProvider.provider
+    val runningOnEmulator = deviceProvider is MockProvider
+}
